Add a number of machine delays
The general idea here is that different sorts of machines might have
different parameters -- for example, there might be a delay slot
between issuing an ALU operation and being able to use it.

This patch adds a number of paramaters that may or may not be used by
the toolchain (hint: they're not :)).